Refactor project structure, update dependencies, and add health endpoint
