+++ tsconfig.json
@@ -4,7 +4,7 @@
     "module": "ES2020",
     "lib": ["ES2020", "DOM"],
     "moduleResolution": "node",
-    "rootDir": "src",
+    "rootDir": "src/server",
     "outDir": "dist",
     "strict": true,
     "esModuleInterop": true,
@@ -12,6 +12,6 @@
     "forceConsistentCasingInFileNames": true,
     "types": ["node", "ws"]
   },
-  "include": ["src/**/*"],
+  "include": ["src/server/**/*.ts"],
   "exclude": ["node_modules"]
 }
